Fixed test suites of the ProductCard component.
