+++ src/components/ProductCard/index.js
@@ -3,21 +3,14 @@ import { StyledCardContainer, StyledImg, StyledCardContent, StyledTitleSpan, Sty
 
 
 export default function ProductCard(props) {
-    const mockCard = {
-        "id": "CnKdjU6CyKakQDGHzNln",
-        "category": "Salgado",
-        "price": "1",
-        "photoUrl": "https://static-images.ifood.com.br/image/upload/f_auto,t_high/pratos/65c38aa8-b094-413d-9a80-ddc256bfcc78/201907031404_66194495.jpg",
-        "name": "Bibsfiha carne",
-        "description": "Esfiha deliciosa, receita secreta do Habibs."
-    }
+
     return (
         <StyledCardContainer>
-            <StyledImg src={mockCard.photoUrl} />
+            <StyledImg src={props.product.photoUrl} />
             <StyledCardContent>
-                <StyledTitleSpan>{mockCard.name}</StyledTitleSpan>
-                <StyledContentSpan>{mockCard.description}</StyledContentSpan>
-                <StyledPriceSpan>R$ 0{mockCard.price},00</StyledPriceSpan>
+                <StyledTitleSpan>{props.product.name}</StyledTitleSpan>
+                <StyledContentSpan>{props.product.description}</StyledContentSpan>
+                <StyledPriceSpan>R$ 0{props.product.price},00</StyledPriceSpan>
             </StyledCardContent>
             <StyledCounter>2</StyledCounter>
             <StyledButton>adicionar</StyledButton>
